Make for loop have exact number of iterations.
Because of floating point rounding errors, the last increment of step might
exceed ToDegree by a tiny fraction. In this case the loop would miss the
last iteration. Adding 1E-3 to ToDegree is sufficient to suppress this case,
since the minimum value of ByDegrees is 0.1.
The for loop is converted to a while loop to avoid clang-tidy warning about
floating-point loop variable.