Make for loop have exact number of iterations.
The intent is that the loop has floor(Wd) iterations, where the last iteration
is Wn exactly.
Because of floating point rounding errors, the last increment of W might
exceed Wn by a tiny fraction. In this case the loop would miss the
last iteration. Adding 1E-3 to Wn suppresses this case. Since Ws is always at
least 1.0, no extra step can be introduced by the 1E-3.

The for loop is converted to a while loop to avoid clang-tidy warning about
floating-point loop variable.