Make for loop always have 4 iterations.
The intent is that the loop has 4 iterations, where the last iteration is
ToDegree exactly.
Because of floating point rounding errors, the last increment of angle might
exceed toDegree by a tiny fraction. In this case the loop would miss the
last iteration. Adding 1E-3 to ToDegree suppresses this case. Since ByDegrees
is 10.0 / 3.0, no extra step can be introduced by the 1E-3.

The for loop is converted to a while loop to avoid clang-tidy warning about
floating-point loop variable.