Convert for loops with double precision loop variables to while loops.
These loops have a step size that may vary by a factor of 1.5 between calls.
There is no intent that the last loop step should be exactly the value of the
loop end condition.

The for loop is converted to a while loop to avoid clang-tidy warning about
floating-point loop variable.