Ensure that loops have exactly 8 steps.
The intent is that i runs from 0.1 to 0.8 in steps of 0.1.
Due to floating point rounding errors, the 8th increment of i might
miss 0.9 by a tiny fraction. In this case the loop would have 9 steps.
Reducing the loop end condition by half a step will suppress this case.

The for loop is converted to a while loop to avoid clang-tidy warning about
floating-point loop variable.