Convert for loops with double precision loop variables to while loops.
These loops have a very variable number of steps, depending on input data.
Since Vend = maxVB + Vstep, the intent is that the last step should be greater
than maxVB. In extreme cases the last two steps may be maxVB and Vend. The
1E-3 suppresses this case.

The for loop is converted to a while loop to avoid clang-tidy warning about
floating-point loop variable.